Separate app and server.
