+++ run.bat
@@ -1,4 +1,4 @@
 @echo off
 set RWKV_JIT_ON=1
 set RWKV_CUDA_ON=0
-python3 server.py
+python3 app.py
