+++ run.sh
@@ -1,3 +1,3 @@
 export RWKV_JIT_ON=1
 export RWKV_CUDA_ON=1
-python3 server.py
+python3 app.py
