+++ server.py
@@ -1,32 +1,8 @@
-from flask import Flask, request
-import os
-import requests
 import chat
 import re
-import datetime
-import logging
-from urllib.parse import quote
-import markdown
-import imgkit
 
 from prompt import User, SCENARIO_ALICE, SCENARIO_ELOISE, SCENARIO_NEURO
 from chat import GenerateMode, CHAT_SAMPLER, INSTRUCT_SAMPLER, model
-
-app = Flask(__name__)
-
-
-formatter = logging.Formatter("%(asctime)s - %(levelname)s - %(message)s")
-handler = logging.FileHandler(f"logs/eloise-{datetime.date.today()}.txt")
-handler.setFormatter(formatter)
-
-logger = logging.getLogger("eloise")
-logger.addHandler(handler)
-logger.setLevel(logging.INFO)
-
-
-banned_users = []
-banned_groups = []
-non_chat_groups = []
 
 
 try:
@@ -35,9 +11,6 @@ try:
 except:
     print("Please provide your QQ number in `qq.txt`")
     QQ = ""
-
-IMAGE_THRESHOLD = 1024
-IMAGE_WIDTH = 600
 
 CHAT_HELP_COMMAND = "`-c, -chat <text>`"
 PRIVATE_HELP_COMMAND = "`<text>`"
@@ -79,8 +52,6 @@ with open("./help.md", 'r') as file:
         '<inst_ap>', str(INSTRUCT_SAMPLER.presence_penalty))
     HELP_MESSAGE = HELP_MESSAGE.replace(
         '<inst_ar>', str(INSTRUCT_SAMPLER.penalty_range))
-
-received_messages = set()
 
 
 def commands(user: User, message, enable_chat=False, is_private=False):
@@ -156,85 +127,5 @@ def commands(user: User, message, enable_chat=False, is_private=False):
     return matched, prompt, reply
 
 
-@app.route('/', methods=["POST"])
-def handle_post():
-    try:
-        json = request.get_json()
-        type = json['message_type']
-        message = json['raw_message']
-        message_id = json['message_id']
-        sender = json['sender']
-        user = User(sender['user_id'], sender['nickname'], sender['sex'])
-    except:
-        return 'OK'
-
-    if user in banned_users:
-        return 'OK'
-    if message_id in received_messages:
-        return 'OK'
-    if len(received_messages) > 500:
-        received_messages.clear()
-
-    if type == 'private':
-        matched, prompt, reply = commands(
-            user, message, enable_chat=True, is_private=True)
-
-        if matched:
-            logger.info(f"{user.nickname}({user.id}): {prompt}")
-            logger.info(reply)
-            received_messages.add(message_id)
-            if len(reply) > IMAGE_THRESHOLD or reply.count('\n') > 2:
-                options = {'font-family': 'SimSun'}
-                html = markdown.markdown(
-                    reply, extensions=['extra', 'nl2br', 'sane_lists', 'codehilite'], options=options)
-
-                file = f"./images/{user.id} {datetime.datetime.now().isoformat()}.png"
-                file = file.replace(' ', '-')
-                path = os.path.abspath(file)
-                options = {'width': IMAGE_WIDTH}
-                imgkit.from_string(
-                    html, file, css='styles.css', options=options)
-                requests.get(
-                    f"http://127.0.0.1:5700/send_private_msg?user_id={user.id}&message=[CQ:image,file=file:///{path}]")
-            else:
-                requests.get(
-                    f"http://127.0.0.1:5700/send_private_msg?user_id={user.id}&message={quote(reply)}")
-    elif type == 'group':
-        try:
-            group_id = int(json['group_id'])
-        except:
-            return 'OK'
-        if group_id in banned_groups:
-            return 'OK'
-        enable_chat = group_id not in non_chat_groups
-
-        matched, prompt, reply = commands(
-            user, message, enable_chat, is_private=False)
-        if matched:
-            logger.info(f"{group_id}: {user.nickname}({user.id}): {prompt}")
-            logger.info(reply)
-            received_messages.add(message_id)
-            if len(reply) > IMAGE_THRESHOLD or reply.count('\n') > 2:
-                options = {'font-family': 'SimSun'}
-                html = markdown.markdown(
-                    reply, extensions=['extra', 'nl2br', 'sane_lists', 'codehilite'], options=options)
-
-                file = f"./images/{user.id} {datetime.datetime.now().isoformat()}.png"
-                file = file.replace(' ', '-')
-                path = os.path.abspath(file)
-                options = {'width': IMAGE_WIDTH}
-                imgkit.from_string(
-                    html, file, css='styles.css', options=options)
-                requests.get(
-                    f"http://127.0.0.1:5700/send_group_msg?group_id={group_id}&message=[CQ:reply,id={message_id}][CQ:image,file=file:///{path}]")
-            else:
-                requests.get(
-                    f"http://127.0.0.1:5700/send_group_msg?group_id={group_id}&message=[CQ:reply,id={message_id}]{quote(reply)}")
-
-    return 'OK'
-
-
-if __name__ == '__main__':
-    print("Starting server...")
+def init():
     chat.init_run()
-    app.run(debug=False, host='127.0.0.1', port=8000, threaded=False)
